Separate finite element field evaluation class
