+++ core/source/cmake/finite_element_source_files.cmake
@@ -16,6 +16,7 @@ SET( FINITE_ELEMENT_CORE_SRCS
 	source/finite_element/finite_element_basis.cpp
 	source/finite_element/finite_element_discretization.cpp
 	source/finite_element/finite_element_field.cpp
+	source/finite_element/finite_element_field_evaluation.cpp
 	source/finite_element/finite_element_mesh.cpp
 	source/finite_element/finite_element_nodeset.cpp
 	source/finite_element/finite_element_region.cpp
@@ -31,6 +32,7 @@ SET( FINITE_ELEMENT_CORE_HDRS
 	source/finite_element/finite_element_basis.h
 	source/finite_element/finite_element_discretization.h
 	source/finite_element/finite_element_field.hpp
+	source/finite_element/finite_element_field_evaluation.hpp
 	source/finite_element/finite_element_field_private.hpp
 	source/finite_element/finite_element_mesh.hpp
 	source/finite_element/finite_element_nodeset.hpp
