+++ core/source/finite_element/finite_element.cpp
@@ -9,9 +9,6 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-/*???DB.  Testing */
-#define DOUBLE_FOR_DOT_PRODUCT
-
 #include <cstddef>
 #include <cstdlib>
 #include <cstdio>
@@ -23,6 +20,7 @@
 #include "general/indexed_list_stl_private.hpp"
 #include <math.h>
 #include "finite_element/finite_element.h"
+#include "finite_element/finite_element_field_evaluation.hpp"
 #include "finite_element/finite_element_field_private.hpp"
 #include "finite_element/finite_element_mesh.hpp"
 #include "finite_element/finite_element_nodeset.hpp"
@@ -132,77 +130,6 @@ DESCRIPTION :
 
 }; /* struct FE_node */
 
-struct FE_element_field_values
-/*******************************************************************************
-LAST MODIFIED : 1 May 2003
-
-DESCRIPTION :
-The values need to calculate a field on an element.  These structures are
-calculated from the element field as required and are then destroyed.
-==============================================================================*/
-{
-	/* the field these values are for */
-	struct FE_field *field;
-	/* the element these values are for */
-	struct FE_element *element;
-	/* the element the field was inherited from */
-	struct FE_element *field_element;
-	/* whether or not these values depend on time */
-	int time_dependent;
-	/* if the values are time dependent, the time at which they were calculated */
-	FE_value time;
-	/* number of sub-elements in each xi-direction of element for each component.
-	 * If NULL (for a component) then field is not grid based for that component.
-		 Notes:
-		1.  Grid sub-elements are linear in each direction.  This means that
-			<component_number_of_values> is not used
-		2.  the grid-point values are not blended (to monomial) and so
-			<component_standard_basis_functions> and
-			<component_standard_basis_function_arguments> are not used
-		3.  for grid-based <destroy_standard_basis_arguments> is used to specify
-			if the <component_values> should be destroyed (element field has been
-			inherited) */
-	int **component_number_in_xi;
-	/* a flag to specify whether or not values have also been calculated for the
-		derivatives of the field with respect to the xi coordinates */
-	char derivatives_calculated;
-	/* a flag added to specify if the element field component modify function is
-		ignored */
-	/* specify whether the standard basis arguments should be destroyed (element
-		field has been inherited) or not be destroyed (element field is defined for
-		the element and the basis arguments are being used) */
-	char destroy_standard_basis_arguments;
-	/* the number of field components */
-	int number_of_components;
-	/* the number of values for each component */
-	int *component_number_of_values;
-	/* the values_storage for each component if grid-based */
-	const Value_storage **component_grid_values_storage;
-	/* grid_offset_in_xi is allocated with 2^number_of_xi_coordinates integers
-		 giving the increment in index into the values stored with the top_level
-		 element for the grid. For top_level_elements the first value is 1, the
-		 second is (number_in_xi[0]+1), the third is
-		 (number_in_xi[0]+1)*(number_in_xi[1]+1) etc. The base_grid_offset is 0 for
-		 top_level_elements. For faces and lines these values are adjusted to get
-		 the correct index for the top_level_element */
-	int *component_base_grid_offset,**component_grid_offset_in_xi;
-	/* following allocated with 2^number_of_xi for grid-based fields for use in
-		 calculate_FE_element_field */
-	int *element_value_offsets;
-	/* the values for each component */
-	FE_value **component_values;
-	/* the standard basis function for each component */
-	Standard_basis_function **component_standard_basis_functions;
-	/* the arguments for the standard basis function for each component */
-	int **component_standard_basis_function_arguments;
-	/* working space for evaluating basis */
-	FE_value *basis_function_values;
-
-	int access_count;
-}; /* struct FE_element_field_values */
-
-FULL_DECLARE_INDEXED_LIST_TYPE(FE_element_field_values);
-
 /**
  * @see struct FE_element_type_node_sequence.
  */
@@ -227,7 +154,7 @@ uses them to find faces and lines for elements without them, if they exist.
 ???RC Can only match faces correctly for coordinate fields with standard node
 to element maps and no versions. A grid-based coordinate field would fail
 utterly since it has no nodes. A possible future solution for all cases is to
-match the geometry exactly either by using the FE_element_field_values
+match the geometry exactly either by using the FE_element_field_evaluation
 (coefficients of the monomial basis functions), although there is a problem with
 xi-directions not matching up, or actual centre positions of the face being a
 trivial rejection, narrowing down to a single face or list of faces to compare
@@ -5728,73 +5655,6 @@ Sets values appropriately if element_dimension = top_level_element_dimension.
 	return (return_code);
 } /* calculate_grid_field_offsets */
 
-int FE_element_field_values_are_for_element_and_time(
-	struct FE_element_field_values *element_field_values,
-	struct FE_element *element,FE_value time,struct FE_element *field_element)
-/*******************************************************************************
-LAST MODIFIED : 3 December 2001
-
-DESCRIPTION :
-Returns true if the <element_field_values> originated from <element>, either
-directly or inherited from <field_element>. If <field_element> is NULL no match
-is required with the field_element in the <element_field_values>.
-==============================================================================*/
-{
-	int return_code;
-
-	ENTER(FE_element_field_values_are_for_element_and_time);
-	if (element_field_values&&element)
-	{
-		return_code=(element_field_values->element==element)&&
-			((!field_element)||(element_field_values->field_element==field_element))
-			&&((!element_field_values->time_dependent)||
-		   (element_field_values->time==time));
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"FE_element_field_values_are_for_element_and_time.  Invalid argument(s)");
-		return_code=0;
-	}
-	LEAVE;
-
-	return (return_code);
-} /* FE_element_field_values_are_for_element_and_time */
-
-int FE_element_field_values_have_derivatives_calculated(
-	struct FE_element_field_values *element_field_values)
-/*******************************************************************************
-LAST MODIFIED : 10 March 2003
-
-DESCRIPTION :
-Returns true if the <element_field_values> are valid for calculating derivatives.
-==============================================================================*/
-{
-	int return_code;
-
-	ENTER(FE_element_field_values_have_derivatives_calculated);
-	if (element_field_values)
-	{
-		if (element_field_values->derivatives_calculated)
-		{
-			return_code = 1;
-		}
-		else
-		{
-			return_code = 0;
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"FE_element_field_values_have_derivatives_calculated.  Invalid argument(s)");
-		return_code = 0;
-	}
-	LEAVE;
-
-	return (return_code);
-} /* FE_element_field_values_have_derivatives_calculated */
-
 PROTOTYPE_ACCESS_OBJECT_FUNCTION(FE_element)
 {
 	if (object)
@@ -5892,1083 +5752,6 @@ int adjacent_FE_element(struct FE_element *element,
 	}
 	return return_code;
 }
-
-struct FE_element_field_values *CREATE(FE_element_field_values)(void)
-/*******************************************************************************
-LAST MODIFIED : 1 May 2003
-
-DESCRIPTION :
-Creates a blank struct FE_element_field_values which can be passed to
-calculate_FE_element_field_values. The structure can be restored to its
-blank state by calling clear_FE_element_field_values; this should be done
-before calling calculate_FE_element_field_values again, and if leaving the
-structure unused for some time so it is not accessing objects.
-==============================================================================*/
-{
-	struct FE_element_field_values *element_field_values;
-
-	ENTER(CREATE(FE_element_field_values));
-	if (ALLOCATE(element_field_values, struct FE_element_field_values, 1))
-	{
-		/* clear the contents of the structure */
-		element_field_values->field = (struct FE_field *)NULL;
-		element_field_values->element = (struct FE_element *)NULL;
-		element_field_values->field_element = (struct FE_element *)NULL;
-		element_field_values->time_dependent = 0;
-		element_field_values->time = 0.0;
-		element_field_values->component_number_in_xi = (int **)NULL;
-		element_field_values->derivatives_calculated = 0;
-		element_field_values->destroy_standard_basis_arguments = 0;
-		element_field_values->number_of_components = 0;
-		element_field_values->component_number_of_values = (int *)NULL;
-		element_field_values->component_grid_values_storage =
-			(const Value_storage **)NULL;
-		element_field_values->component_base_grid_offset = (int *)NULL;
-		element_field_values->component_grid_offset_in_xi = (int **)NULL;
-		element_field_values->element_value_offsets = (int *)NULL;
-		element_field_values->component_values = (FE_value **)NULL;
-		element_field_values->component_standard_basis_functions =
-			(Standard_basis_function **)NULL;
-		element_field_values->component_standard_basis_function_arguments =
-			(int **)NULL;
-		element_field_values->basis_function_values = (FE_value *)NULL;
-		element_field_values->access_count = 0;
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"CREATE(FE_element_field_values).  Could not allocate structure");
-	}
-	LEAVE;
-
-	return (element_field_values);
-} /* CREATE(FE_element_field_values) */
-
-int calculate_FE_element_field_values(cmzn_element *element,
-	struct FE_field *field, FE_value time, char calculate_derivatives,
-	struct FE_element_field_values *element_field_values,
-	cmzn_element *topLevelElement)
-{
-	FE_value *blending_matrix,
-		*derivative_value,*inherited_value,*inherited_values,scalar,
-		*second_derivative_value,*transformation,*value,**values_address;
-	int cn,**component_number_in_xi,
-		*component_base_grid_offset, *element_value_offsets, i,
-		j,k,grid_maximum_number_of_values, maximum_number_of_values,
-		number_of_grid_based_components,
-		number_of_inherited_values,number_of_polygon_verticies,number_of_values,
-		*number_of_values_address,offset,order,*orders,polygon_offset,power,
-		row_size,**standard_basis_arguments_address;
-	Standard_basis_function **standard_basis_address;
-	// this had used DOUBLE_FOR_DOT_PRODUCT, but FE_value will be at least double precision now
-	FE_value sum;
-
-	if (!((element) && (field) && (element_field_values)))
-	{
-		display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  Invalid argument(s)");
-		return 0;
-	}
-	FE_value coordinate_transformation[MAXIMUM_ELEMENT_XI_DIMENSIONS*MAXIMUM_ELEMENT_XI_DIMENSIONS];
-	cmzn_element *fieldElement = field->getOrInheritOnElement(element,
-		/*inherit_face_number*/-1, topLevelElement, coordinate_transformation);
-	if (!fieldElement)
-	{
-		display_message(ERROR_MESSAGE,
-			"calculate_FE_element_field_values.  Field %s not defined on %d-D element %d",
-			field->name, element->getDimension(), element->getIdentifier());
-		return 0;
-	}
-	const FE_mesh *mesh = element->getMesh();
-	if (!mesh)
-	{
-		display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  Invalid element");
-		return 0;
-	}
-	const FE_nodeset *nodeset = mesh->getNodeset();
-	if (!nodeset)
-	{
-		display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  No nodeset, invalid mesh");
-		return 0;
-	}
-
-	int return_code = 1;
-	const int elementDimension = element->getDimension();
-	const int fieldElementDimension = fieldElement->getDimension();
-	const int number_of_components = field->number_of_components;
-	switch (field->fe_field_type)
-	{
-		case CONSTANT_FE_FIELD:
-		{
-			/* constant fields do not use the values except to remember the
-					element and field they are for */
-			element_field_values->field=ACCESS(FE_field)(field);
-			element_field_values->element = element->access();
-			/* store fieldElement since we are now able to suggest through the
-					topLevelElement clue which one we get. Must compare element
-					and fieldElement to ensure field values are still valid for
-					a given line or face. */
-			element_field_values->field_element = fieldElement->access();
-			element_field_values->component_number_in_xi=(int **)NULL;
-			/* derivatives will be calculated in calculate_FE_element_field */
-			/*???DB.  Assuming linear */
-			element_field_values->derivatives_calculated=1;
-			element_field_values->destroy_standard_basis_arguments=0;
-			element_field_values->number_of_components=number_of_components;
-			element_field_values->component_number_of_values=(int *)NULL;
-			element_field_values->component_grid_values_storage=
-				(const Value_storage **)NULL;
-			element_field_values->component_base_grid_offset=(int *)NULL;
-			element_field_values->component_grid_offset_in_xi=(int **)NULL;
-			element_field_values->element_value_offsets=(int *)NULL;
-			/* clear arrays not used for grid-based fields */
-			element_field_values->component_values=(FE_value **)NULL;
-			element_field_values->component_standard_basis_functions=
-				(Standard_basis_function **)NULL;
-			element_field_values->component_standard_basis_function_arguments=
-				(int **)NULL;
-			element_field_values->basis_function_values=(FE_value *)NULL;
-			element_field_values->time_dependent = 0;
-			element_field_values->time = time;
-		} break;
-		case INDEXED_FE_FIELD:
-		{
-			if (calculate_FE_element_field_values(element,field->indexer_field,
-				time,calculate_derivatives,element_field_values,topLevelElement))
-			{
-				/* restore pointer to original field - has the indexer_field in
-						it anyway */
-				REACCESS(FE_field)(&(element_field_values->field),field);
-			}
-			else
-			{
-				display_message(ERROR_MESSAGE,"calculate_FE_element_field_values.  "
-					"Cannot calculate element field values for indexer field");
-				return_code=0;
-			}
-		} break;
-		case GENERAL_FE_FIELD:
-		{
-			ALLOCATE(number_of_values_address,int,number_of_components);
-			ALLOCATE(values_address,FE_value *,number_of_components);
-			ALLOCATE(standard_basis_address,Standard_basis_function *,
-				number_of_components);
-			ALLOCATE(standard_basis_arguments_address,int *,
-				number_of_components);
-			blending_matrix=(FE_value *)NULL;
-			ALLOCATE(component_number_in_xi, int *, number_of_components);
-			if (number_of_values_address&&values_address&&
-				standard_basis_address&&standard_basis_arguments_address&&
-				component_number_in_xi)
-			{
-				for (i=0;i<number_of_components;i++)
-				{
-					number_of_values_address[i] = 0;
-					values_address[i] = (FE_value *)NULL;
-					standard_basis_address[i] = (Standard_basis_function *)NULL;
-					standard_basis_arguments_address[i]=(int *)NULL;
-					/* following is non-NULL only for grid-based components */
-					component_number_in_xi[i] = NULL;
-				}
-				element_field_values->component_number_in_xi = component_number_in_xi;
-				element_field_values->field = field->access();
-				element_field_values->element = element->access();
-				element_field_values->field_element = fieldElement->access();
-				element_field_values->time_dependent =
-					FE_field_has_multiple_times(element_field_values->field);
-				element_field_values->time = time;
-				element_field_values->derivatives_calculated=calculate_derivatives;
-				if (fieldElementDimension > elementDimension)
-				{
-					element_field_values->destroy_standard_basis_arguments=1;
-				}
-				else
-				{
-					element_field_values->destroy_standard_basis_arguments=0;
-				}
-				element_field_values->number_of_components=number_of_components;
-				element_field_values->component_number_of_values=
-					number_of_values_address;
-				/* clear arrays only used for grid-based fields */
-				element_field_values->component_grid_values_storage=
-					(const Value_storage **)NULL;
-				element_field_values->component_grid_offset_in_xi=(int **)NULL;
-				element_field_values->element_value_offsets=(int *)NULL;
-
-				element_field_values->component_values=values_address;
-				element_field_values->component_standard_basis_functions=
-					standard_basis_address;
-				element_field_values->component_standard_basis_function_arguments=
-					standard_basis_arguments_address;
-				/* maximum_number_of_values starts off big enough for linear grid with derivatives */
-				grid_maximum_number_of_values=elementDimension+1;
-				for (i=elementDimension;i>0;i--)
-				{
-					grid_maximum_number_of_values *= 2;
-				}
-				maximum_number_of_values = grid_maximum_number_of_values;
-
-				const FE_mesh_field_data *meshFieldData = field->meshFieldData[fieldElementDimension - 1];
-				FE_basis *previous_basis = 0;
-				return_code=1;
-				number_of_grid_based_components = 0;
-				int **component_grid_offset_in_xi = 0;
-				const DsLabelIndex fieldElementIndex = fieldElement->getIndex();
-				for (int component_number = 0; return_code && (component_number < number_of_components); ++component_number)
-				{
-					const FE_element_field_template *componentEFT =
-						meshFieldData->getComponentMeshfieldtemplate(component_number)->getElementfieldtemplate(fieldElementIndex);
-					if (!componentEFT)
-					{
-						return_code = 0;
-						break;
-					}
-					if ((componentEFT->getParameterMappingMode() == CMZN_ELEMENTFIELDTEMPLATE_PARAMETER_MAPPING_MODE_ELEMENT)
-						&& (0 != componentEFT->getLegacyGridNumberInXi()))
-					{
-						const int *top_level_component_number_in_xi = componentEFT->getLegacyGridNumberInXi();
-						if (!top_level_component_number_in_xi)
-						{
-							display_message(ERROR_MESSAGE,
-								"calculate_FE_element_field_values.  Element parameter mapping only implemented for legacy grid");
-							return_code = 0;
-							break;
-						}
-						++number_of_grid_based_components;
-						if (number_of_grid_based_components == number_of_components)
-						{
-							element_field_values->derivatives_calculated=1;
-						}
-						/* one-off allocation of arrays only needed for grid-based components */
-						if (NULL == element_field_values->component_grid_values_storage)
-						{
-							const Value_storage **component_grid_values_storage;
-							ALLOCATE(component_grid_values_storage, const Value_storage *, number_of_components);
-							ALLOCATE(component_base_grid_offset, int, number_of_components);
-							ALLOCATE(component_grid_offset_in_xi, int*, number_of_components);
-							ALLOCATE(element_value_offsets, int, grid_maximum_number_of_values);
-							if (component_grid_values_storage && component_base_grid_offset &&
-								component_grid_offset_in_xi && element_value_offsets)
-							{
-								for (cn = 0; (cn < number_of_components); cn++)
-								{
-									component_grid_values_storage[cn]=NULL;
-									component_base_grid_offset[cn]=0;
-									component_grid_offset_in_xi[cn]=NULL;
-								}
-								element_field_values->component_grid_values_storage = component_grid_values_storage;
-								element_field_values->component_base_grid_offset = component_base_grid_offset;
-								element_field_values->component_grid_offset_in_xi = component_grid_offset_in_xi;
-								element_field_values->element_value_offsets = element_value_offsets;
-							}
-							else
-							{
-								return_code = 0;
-								break;
-							}
-						}
-						// GRC risky to cache pointers into per-element data
-						FE_mesh_field_data::ComponentBase *componentBase = meshFieldData->getComponentBase(component_number);
-						const int valuesCount = componentEFT->getNumberOfElementDOFs();
-						// following could be done as a virtual function
-						switch (field->value_type)
-						{
-						case FE_VALUE_VALUE:
-							{
-								auto component = static_cast<FE_mesh_field_data::Component<FE_value>*>(componentBase);
-								element_field_values->component_grid_values_storage[component_number] =
-									reinterpret_cast<const Value_storage*>(component->getElementValues(fieldElementIndex, valuesCount));
-							} break;
-						case INT_VALUE:
-							{
-								auto component = static_cast<FE_mesh_field_data::Component<int>*>(componentBase);
-								element_field_values->component_grid_values_storage[component_number] =
-									reinterpret_cast<const Value_storage*>(component->getElementValues(fieldElementIndex, valuesCount));
-							} break;
-						default:
-							{
-								display_message(ERROR_MESSAGE,
-									"calculate_FE_element_field_values.  Invalid value type for grid field");
-								return_code = 0;
-							} break;
-						}
-						ALLOCATE(component_number_in_xi[component_number], int, elementDimension);
-						ALLOCATE(component_grid_offset_in_xi[component_number], int, elementDimension);
-						if ((NULL != component_number_in_xi[component_number]) &&
-							(NULL != component_grid_offset_in_xi[component_number]))
-						{
-							element_field_values->component_base_grid_offset[component_number] = 0;
-							if (!calculate_grid_field_offsets(elementDimension,
-								fieldElementDimension, top_level_component_number_in_xi,
-								coordinate_transformation, component_number_in_xi[component_number],
-								&(element_field_values->component_base_grid_offset[component_number]),
-								component_grid_offset_in_xi[component_number]))
-							{
-								display_message(ERROR_MESSAGE,
-									"calculate_FE_element_field_values.  "
-									"Could not calculate grid field offsets");
-								return_code=0;
-								break;
-							}
-						}
-						else
-						{
-							return_code = 0;
-							break;
-						}
-					}
-					else /* not grid-based; includes non-grid element-based */
-					{
-						// calculate element values for component on the fieldElement
-						const int basisFunctionCount = componentEFT->getNumberOfFunctions();
-						ALLOCATE(*values_address, FE_value, basisFunctionCount);
-						if (!(*values_address))
-						{
-							return_code = 0;
-							break;
-						}
-						switch (componentEFT->getParameterMappingMode())
-						{
-						case CMZN_ELEMENTFIELDTEMPLATE_PARAMETER_MAPPING_MODE_NODE:
-						{
-							if (0 == (*number_of_values_address = global_to_element_map_values(field, component_number,
-								componentEFT, fieldElement, time, nodeset, *values_address)))
-							{
-								display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  "
-									"Could not calculate node-based values for field %s in %d-D element %d",
-									field->name, fieldElementDimension, fieldElement->getIdentifier());
-								return_code = 0;
-							}
-						} break;
-						case CMZN_ELEMENTFIELDTEMPLATE_PARAMETER_MAPPING_MODE_ELEMENT:
-						{
-							// element-based mapping stores the parameters ready for use in the element
-							if (field->value_type != FE_VALUE_VALUE)
-							{
-								display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  "
-									"Element-based non-grid field %s only implemented for real values", field->name);
-								return_code = 0;
-								break;
-							}
-							auto component = static_cast<FE_mesh_field_data::Component<FE_value> *>(meshFieldData->getComponentBase(component_number));
-							const FE_value *values = component->getElementValues(fieldElementIndex, basisFunctionCount);
-							if (!values)
-							{
-								display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  "
-									"Element-based field %s has no values at %d-D element %d",
-									field->name, fieldElementDimension, fieldElement->getIdentifier());
-								return_code = 0;
-								break;
-							}
-							memcpy(*values_address, values, basisFunctionCount*sizeof(FE_value));
-							*number_of_values_address = basisFunctionCount;
-						} break;
-						case CMZN_ELEMENTFIELDTEMPLATE_PARAMETER_MAPPING_MODE_FIELD:
-						{
-							if (field->value_type != FE_VALUE_VALUE)
-							{
-								display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  "
-									"Field-based field %s only implemented for real values", field->name);
-								return_code = 0;
-								break;
-							}
-							if (!get_FE_field_FE_value_value(field, component_number, *values_address))
-							{
-								display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  "
-									"Field-based field %s has no values at %d-D element %d",
-									field->name, fieldElementDimension, fieldElement->getIdentifier());
-								return_code = 0;
-								break;
-							}
-							*number_of_values_address = 1;
-						} break;
-						case CMZN_ELEMENTFIELDTEMPLATE_PARAMETER_MAPPING_MODE_INVALID:
-						{
-							display_message(ERROR_MESSAGE, "calculate_FE_element_field_values.  "
-								"Invalid parameter mapping mode for field %s in %d-D element %d",
-								field->name, fieldElementDimension, fieldElement->getIdentifier());
-							return_code = 0;
-						} break;
-						}
-						if (!return_code)
-							break;
-						if (previous_basis == componentEFT->getBasis())
-						{
-							*standard_basis_address= *(standard_basis_address-1);
-							*standard_basis_arguments_address=
-								*(standard_basis_arguments_address-1);
-						}
-						else
-						{
-							previous_basis = componentEFT->getBasis();
-							if (blending_matrix)
-							{
-								/* SAB We don't want to keep the old one */
-								DEALLOCATE(blending_matrix);
-								blending_matrix=NULL;
-							}
-							*standard_basis_address = FE_basis_get_standard_basis_function(previous_basis);
-							if (fieldElementDimension > elementDimension)
-							{
-								return_code=calculate_standard_basis_transformation(
-									previous_basis,coordinate_transformation,
-									elementDimension,standard_basis_arguments_address,
-									&number_of_inherited_values,standard_basis_address,
-									&blending_matrix);
-							}
-							else
-							{
-								/* standard basis transformation is just a big identity matrix, so don't compute */
-								/* also use the real basis arguments */
-								*standard_basis_arguments_address =
-									const_cast<int *>(FE_basis_get_standard_basis_function_arguments(previous_basis));
-							}
-						}
-						if (return_code)
-						{
-							if (fieldElement == element)
-							{
-								/* values already correct regardless of basis, but must make space for derivatives if needed */
-								if (calculate_derivatives)
-								{
-									if (REALLOCATE(inherited_values,*values_address,FE_value,
-										(elementDimension+1)*(*number_of_values_address)))
-									{
-										*values_address=inherited_values;
-									}
-									else
-									{
-										display_message(ERROR_MESSAGE,
-											"calculate_FE_element_field_values.  Could not reallocate values");
-										return_code=0;
-									}
-								}
-							}
-							else if ((monomial_basis_functions== *standard_basis_address)||
-								(polygon_basis_functions== *standard_basis_address))
-							{
-								/* project the fieldElement values onto the lower-dimension element
-										using the affine transformation */
-								/* allocate memory for the element values */
-								if (calculate_derivatives)
-								{
-									ALLOCATE(inherited_values,FE_value,
-										(elementDimension+1)*number_of_inherited_values);
-								}
-								else
-								{
-									ALLOCATE(inherited_values,FE_value,
-										number_of_inherited_values);
-								}
-								if (inherited_values)
-								{
-									row_size= *number_of_values_address;
-									inherited_value=inherited_values;
-									for (j=0;j<number_of_inherited_values;j++)
-									{
-										sum=0;
-										value= *values_address;
-										transformation=blending_matrix+j;
-										for (i=row_size;i>0;i--)
-										{
-											sum += (*transformation)*(*value);
-											value++;
-											transformation += number_of_inherited_values;
-										}
-										*inherited_value=(FE_value)sum;
-										inherited_value++;
-									}
-									DEALLOCATE(*values_address);
-									*values_address=inherited_values;
-									*number_of_values_address=number_of_inherited_values;
-								}
-								else
-								{
-									display_message(ERROR_MESSAGE,
-										"calculate_FE_element_field_values.  "
-										"Insufficient memory for inherited_values");
-									DEALLOCATE(*values_address);
-									return_code=0;
-								}
-							}
-							else
-							{
-								display_message(ERROR_MESSAGE,
-									"calculate_FE_element_field_values.  Invalid basis");
-								return_code=0;
-							}
-							if (return_code)
-							{
-								if (calculate_derivatives)
-								{
-									/* calculate the derivatives with respect to the xi
-											coordinates */
-									if (monomial_basis_functions==
-										*standard_basis_address)
-									{
-										number_of_values= *number_of_values_address;
-										value= *values_address;
-										derivative_value=value+number_of_values;
-										orders= *standard_basis_arguments_address;
-										offset=1;
-										for (i=elementDimension;i>0;i--)
-										{
-											orders++;
-											order= *orders;
-											for (j=0;j<number_of_values;j++)
-											{
-												/* calculate derivative value */
-												power=(j/offset)%(order+1);
-												if (order==power)
-												{
-													*derivative_value=0;
-												}
-												else
-												{
-													*derivative_value=
-														(FE_value)(power+1)*value[j+offset];
-												}
-												/* move to the next derivative value */
-												derivative_value++;
-											}
-											offset *= (order+1);
-										}
-									}
-									else if (polygon_basis_functions==
-										*standard_basis_address)
-									{
-										number_of_values= *number_of_values_address;
-										value= *values_address;
-										derivative_value=value+number_of_values;
-										orders= *standard_basis_arguments_address;
-										offset=1;
-										for (i=elementDimension;i>0;i--)
-										{
-											orders++;
-											order= *orders;
-											if (order<0)
-											{
-												/* polygon */
-												order= -order;
-												if (order%2)
-												{
-													/* calculate derivatives with respect to
-														both polygon coordinates */
-													order /= 2;
-													polygon_offset=order%elementDimension;
-													order /= elementDimension;
-													number_of_polygon_verticies=
-														(-orders[polygon_offset])/2;
-													/* first polygon coordinate is
-														circumferential */
-													second_derivative_value=derivative_value+
-														(polygon_offset*number_of_values);
-													order=4*number_of_polygon_verticies;
-													scalar=
-														(FE_value)number_of_polygon_verticies;
-													for (j=0;j<number_of_values;j++)
-													{
-														/* calculate derivative values */
-														k=(j/offset)%order;
-														switch (k/number_of_polygon_verticies)
-														{
-															case 0:
-															{
-																*derivative_value=scalar*value[j+
-																	number_of_polygon_verticies*offset];
-																*second_derivative_value=value[j+
-																	2*number_of_polygon_verticies*
-																	offset];
-															} break;
-															case 1:
-															{
-																*derivative_value=0;
-																*second_derivative_value=value[j+
-																	2*number_of_polygon_verticies*
-																	offset];
-															} break;
-															case 2:
-															{
-																*derivative_value=scalar*value[j+
-																	number_of_polygon_verticies*offset];
-																*second_derivative_value=0;
-															} break;
-															case 3:
-															{
-																*derivative_value=0;
-																*second_derivative_value=0;
-															} break;
-														}
-														/* move to the next derivative value */
-														derivative_value++;
-														second_derivative_value++;
-													}
-													offset *= order;
-												}
-												else
-												{
-													/* second polgon xi.  Derivatives already
-														calculated */
-													derivative_value += number_of_values;
-												}
-											}
-											else
-											{
-												/* not polygon */
-												for (j=0;j<number_of_values;j++)
-												{
-													/* calculate derivative value */
-													power=(j/offset)%(order+1);
-													if (order==power)
-													{
-														*derivative_value=0;
-													}
-													else
-													{
-														*derivative_value=
-															(FE_value)(power+1)*value[j+offset];
-													}
-													/* move to the next derivative value */
-													derivative_value++;
-												}
-												offset *= (order+1);
-											}
-										}
-									}
-									else
-									{
-										display_message(ERROR_MESSAGE,
-											"calculate_FE_element_field_values.  "
-											"Invalid basis");
-										DEALLOCATE(*values_address);
-										return_code=0;
-									}
-								}
-							}
-						}
-						if (*number_of_values_address>maximum_number_of_values)
-						{
-							maximum_number_of_values= *number_of_values_address;
-						}
-						number_of_values_address++;
-						values_address++;
-						standard_basis_address++;
-						standard_basis_arguments_address++;
-					}
-				}
-				if (return_code)
-				{
-					if (!((maximum_number_of_values>0)&&
-						(ALLOCATE(element_field_values->basis_function_values,
-						FE_value,maximum_number_of_values))))
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_values.  "
-							"Could not allocate basis_function_values");
-						return_code=0;
-					}
-				}
-			}
-			else
-			{
-				display_message(ERROR_MESSAGE,
-					"calculate_FE_element_field_values.  Insufficient memory");
-				DEALLOCATE(number_of_values_address);
-				DEALLOCATE(values_address);
-				DEALLOCATE(standard_basis_address);
-				DEALLOCATE(standard_basis_arguments_address);
-				return_code=0;
-			}
-			if (blending_matrix)
-			{
-				DEALLOCATE(blending_matrix);
-			}
-		} break;
-		default:
-		{
-			display_message(ERROR_MESSAGE,
-				"calculate_FE_element_field_values.  Unknown field type");
-			return_code=0;
-		} break;
-	} /* switch (field->fe_field_type) */
-	return (return_code);
-}
-
-int FE_element_field_values_differentiate(
-	struct FE_element_field_values *element_field_values, int xi_index)
-/*******************************************************************************
-LAST MODIFIED : 9 May 2007
-
-DESCRIPTION :
-Modifies the calculated values for an FE_field so that it will calculate
-derivatives wrt xi_index for the original field.  The <element_field_values>
-must have already been calculated.  Currently only implemented for monomials.
-==============================================================================*/
-{
-	FE_value *derivative_value, *value;
-	int elementDimension, i, j, k, number_of_values, offset, order,
-		*orders, power, return_code;
-
-	ENTER(FE_element_field_values_differentiate);
-	if (element_field_values && element_field_values->derivatives_calculated)
-	{
-		return_code = 1;
-		elementDimension = element_field_values->element->getDimension();
-		for (k = 0 ; k < element_field_values->number_of_components ; k++)
-		{
-			if (monomial_basis_functions==
-				element_field_values->component_standard_basis_functions[k])
-			{
-				number_of_values = element_field_values->component_number_of_values[k];
-				value = element_field_values->component_values[k];
-
-				/* Copy the specified derivative back into the values */
-				derivative_value = value + number_of_values * (xi_index + 1);
-				for (j=0;j<number_of_values;j++)
-				{
-					*value = *derivative_value;
-					value++;
-					derivative_value++;
-				}
-
-				/* Now differentiate the values monomial as we did to calculate them above */
-
-				value = element_field_values->component_values[k];
-				derivative_value = value + number_of_values;
-
-				orders= element_field_values->component_standard_basis_function_arguments[k];
-				offset = 1;
-
-				for (i=elementDimension;i>0;i--)
-				{
-					orders++;
-					order= *orders;
-					for (j=0;j<number_of_values;j++)
-					{
-						/* calculate derivative value */
-						power=(j/offset)%(order+1);
-						if (order==power)
-						{
-							*derivative_value=0;
-						}
-						else
-						{
-							*derivative_value=
-								(FE_value)(power+1)*value[j+offset];
-						}
-						/* move to the next derivative value */
-						derivative_value++;
-					}
-					offset *= (order+1);
-				}
-			}
-			else
-			{
-				display_message(ERROR_MESSAGE,
-					"FE_element_field_values_differentiate.  Unsupported basis type");
-				return_code=0;
-				break;
-			}
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"FE_element_field_values_differentiate.  Invalid argument");
-		return_code=0;
-	}
-	LEAVE;
-
-	return (return_code);
-} /* FE_element_field_values_differentiate */
-
-int clear_FE_element_field_values(
-	struct FE_element_field_values *element_field_values)
-/*******************************************************************************
-LAST MODIFIED : 1 May 2003
-
-DESCRIPTION :
-Frees the memory for the fields of the <element_field_values> structure.
-Restores <element_field_values> to the blank state it was created with. This
-function must be called before calling calculate_FE_element_field_values again.
-==============================================================================*/
-{
-	FE_value **component_values;
-	int i, return_code;
-	int **component_standard_basis_function_arguments;
-
-	ENTER(clear_FE_element_field_values);
-	return_code=0;
-	if (element_field_values)
-	{
-		return_code=1;
-		if (element_field_values->component_number_in_xi)
-		{
-			if (element_field_values->field)
-			{
-				for (i = 0; i < element_field_values->number_of_components; i++)
-				{
-					if (element_field_values->component_number_in_xi[i])
-					{
-						DEALLOCATE(element_field_values->component_number_in_xi[i]);
-					}
-				}
-			}
-			DEALLOCATE(element_field_values->component_number_in_xi);
-		}
-		DEACCESS(FE_field)(&(element_field_values->field));
-		DEACCESS(FE_element)(&(element_field_values->element));
-		DEACCESS(FE_element)(&(element_field_values->field_element));
-		if (element_field_values->component_number_of_values)
-		{
-			DEALLOCATE(element_field_values->component_number_of_values);
-		}
-		if (element_field_values->component_grid_values_storage)
-		{
-			DEALLOCATE(element_field_values->component_grid_values_storage);
-		}
-		if (element_field_values->component_base_grid_offset)
-		{
-			DEALLOCATE(element_field_values->component_base_grid_offset);
-		}
-		if (element_field_values->component_grid_offset_in_xi)
-		{
-			for (i = 0; i < element_field_values->number_of_components; i++)
-			{
-				DEALLOCATE(element_field_values->component_grid_offset_in_xi[i]);
-			}
-			DEALLOCATE(element_field_values->component_grid_offset_in_xi);
-		}
-		if (element_field_values->element_value_offsets)
-		{
-			DEALLOCATE(element_field_values->element_value_offsets);
-		}
-		if (element_field_values->component_values)
-		{
-			component_values=element_field_values->component_values;
-			for (i=element_field_values->number_of_components;i>0;i--)
-			{
-				DEALLOCATE(*component_values);
-				component_values++;
-			}
-			DEALLOCATE(element_field_values->component_values);
-		}
-		if (element_field_values->component_standard_basis_function_arguments)
-		{
-			if (element_field_values->destroy_standard_basis_arguments)
-			{
-				component_standard_basis_function_arguments=
-					element_field_values->component_standard_basis_function_arguments;
-				for (i=element_field_values->number_of_components;i>0;i--)
-				{
-					if (*component_standard_basis_function_arguments&&((1==i)||
-						(*component_standard_basis_function_arguments!=
-							component_standard_basis_function_arguments[1])))
-					{
-						DEALLOCATE(*component_standard_basis_function_arguments);
-					}
-					component_standard_basis_function_arguments++;
-				}
-			}
-			DEALLOCATE(element_field_values->
-				component_standard_basis_function_arguments);
-		}
-		if (element_field_values->component_standard_basis_functions)
-		{
-			DEALLOCATE(element_field_values->component_standard_basis_functions);
-		}
-		if (element_field_values->basis_function_values)
-		{
-			DEALLOCATE(element_field_values->basis_function_values);
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"clear_FE_element_field_values.  Invalid argument");
-		return_code=0;
-	}
-	LEAVE;
-
-	return (return_code);
-} /* clear_FE_element_field_values */
-
-int DESTROY(FE_element_field_values)(
-	struct FE_element_field_values **element_field_values_address)
-/*******************************************************************************
-LAST MODIFIED : 4 March 2003
-
-DESCRIPTION :
-Destroys the element_field_values at *<element_field_info_address>. Frees the
-memory for the information and sets <*element_field_info_address> to NULL.
-==============================================================================*/
-{
-	int return_code;
-	struct FE_element_field_values *element_field_values;
-
-	ENTER(DESTROY(FE_element_field_values));
-	if ((element_field_values_address) &&
-		(element_field_values = *element_field_values_address))
-	{
-		clear_FE_element_field_values(element_field_values);
-		DEALLOCATE(*element_field_values_address);
-		*element_field_values_address = (struct FE_element_field_values *)NULL;
-		return_code = 1;
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"DESTROY(FE_element_field_values).  Invalid argument(s)");
-		return_code = 0;
-	}
-	LEAVE;
-
-	return (return_code);
-} /* DESTROY(FE_element_field_values) */
-
-DECLARE_OBJECT_FUNCTIONS(FE_element_field_values)
-
-DECLARE_INDEXED_LIST_MODULE_FUNCTIONS(FE_element_field_values, element, \
-	struct FE_element *, compare_pointer)
-
-DECLARE_INDEXED_LIST_FUNCTIONS(FE_element_field_values)
-
-DECLARE_FIND_BY_IDENTIFIER_IN_INDEXED_LIST_FUNCTION(FE_element_field_values, element, \
-	struct FE_element *, compare_pointer)
-
-int FE_element_field_values_get_component_values(
-	struct FE_element_field_values *element_field_values,int component_number,
-	int *number_of_component_values_address,FE_value **component_values_address)
-/*******************************************************************************
-LAST MODIFIED : 1 June 2004
-
-DESCRIPTION :
-Allocates and returns to <component_values_address> the component values for
-<component_number> in <element_field_values>. The number of values is returned
-in <number_of_component_values>.
-It is up to the calling function to deallocate any returned component values.
-==============================================================================*/
-{
-	int return_code;
-
-	ENTER(FE_element_field_values_get_component_values);
-	return_code=0;
-	if (element_field_values&&(element_field_values->element)&&
-		(0<=component_number)&&
-		(component_number<element_field_values->number_of_components)&&
-		number_of_component_values_address&&component_values_address)
-	{
-		if ((element_field_values->component_number_of_values)&&
-			(0<(*number_of_component_values_address=
-			element_field_values->component_number_of_values[component_number]))&&
-			(element_field_values->component_values)&&
-			element_field_values->component_values[component_number]&&
-			ALLOCATE(*component_values_address,FE_value,
-				*number_of_component_values_address))
-		{
-			memcpy(*component_values_address,
-				element_field_values->component_values[component_number],
-				(*number_of_component_values_address)*sizeof(FE_value));
-			return_code=1;
-		}
-		else
-		{
-			display_message(ERROR_MESSAGE,
-				"FE_element_field_values_get_component_values.  "
-				"Component has no values");
-		}
-	}
-	else
-	{
-		if (element_field_values)
-		{
-			display_message(ERROR_MESSAGE,
-				"FE_element_field_values_get_component_values.  "
-				"Invalid argument(s).  %p %p %d %d %p %p",element_field_values,
-				element_field_values->element,component_number,
-				element_field_values->number_of_components,
-				number_of_component_values_address,component_values_address);
-		}
-		else
-		{
-			display_message(ERROR_MESSAGE,
-				"FE_element_field_values_get_component_values.  "
-				"Invalid argument(s).  %p %d %p %p",element_field_values,
-				component_number,number_of_component_values_address,
-				component_values_address);
-		}
-	}
-	LEAVE;
-
-	return (return_code);
-} /* FE_element_field_values_get_component_values */
-
-int FE_element_field_values_get_monomial_component_info(
-	struct FE_element_field_values *element_field_values,int component_number,
-	int *monomial_info)
-/*******************************************************************************
-LAST MODIFIED : 21 May 2004
-
-DESCRIPTION :
-If <component_number> in the <element_field_values> is monomial, integer values
-describing the monomial basis are returned. The first number is the dimension,
-the following numbers are the order of the monomial in each direction, where
-3=cubic, for example.
-<monomial_info> should point to a block of memory big enough to take
-1 + MAXIMUM_ELEMENT_XI_DIMENSIONS integers.
-==============================================================================*/
-{
-	int i, return_code, *source_monomial_info;
-
-	ENTER(FE_element_field_values_get_monomial_component_info);
-	return_code=0;
-	if (element_field_values&&element_field_values->element&&
-		(0<=component_number)&&
-		(component_number<element_field_values->number_of_components)&&
-		monomial_info)
-	{
-		if ((element_field_values->component_standard_basis_function_arguments)&&
-			(source_monomial_info=element_field_values->
-			component_standard_basis_function_arguments[component_number])&&
-			standard_basis_function_is_monomial(element_field_values->
-			component_standard_basis_functions[component_number],
-			(void *)source_monomial_info))
-		{
-			*monomial_info= *source_monomial_info;
-			for (i=1;i<= *monomial_info;i++)
-			{
-				monomial_info[i]=source_monomial_info[i];
-			}
-			return_code=1;
-		}
-		else
-		{
-			display_message(ERROR_MESSAGE,
-				"FE_element_field_values_get_monomial_component_info.  "
-				"Component is not monomial");
-		}
-	}
-	else
-	{
-		if (element_field_values)
-		{
-			display_message(ERROR_MESSAGE,
-				"FE_element_field_values_get_monomial_component_info.  "
-				"Invalid argument(s).  %p %p %d %d %p",element_field_values,
-				element_field_values->element,component_number,
-				element_field_values->number_of_components,monomial_info);
-		}
-		else
-		{
-			display_message(ERROR_MESSAGE,
-				"FE_element_field_values_get_monomial_component_info.  "
-				"Invalid argument(s).  %p %d %p",element_field_values,component_number,
-				monomial_info);
-		}
-	}
-	LEAVE;
-
-	return (return_code);
-} /* FE_element_field_values_get_monomial_component_info */
 
 int calculate_FE_element_field_nodes(struct FE_element *element,
 	int face_number, struct FE_field *field,
@@ -7212,859 +5995,6 @@ int calculate_FE_element_field_nodes(struct FE_element *element,
 	}
 	return (return_code);
 }
-
-int calculate_FE_element_field(int component_number,
-	struct FE_element_field_values *element_field_values,
-	const FE_value *xi_coordinates, FE_value *values, FE_value *jacobian)
-/*******************************************************************************
-LAST MODIFIED : 5 August 2001
-
-DESCRIPTION :
-Calculates the <values> of the field specified by the <element_field_values> at
-the <xi_coordinates>.  The storage for the <values> should have been allocated
-outside the function.  The <jacobian> will be calculated if it is not NULL (and
-the derivatives values have been calculated).  Only the <component_number>+1
-component will be calculated if 0<=component_number<number of components.  For a
-single component, the value will be put in the first position of <values> and
-the derivatives will start at the first position of <jacobian>.
-==============================================================================*/
-{
-	int cn,comp_no,*component_number_of_values,components_to_calculate,
-		i,j,k,l,m, *last_number_in_xi,*number_in_xi,
-		number_of_xi_coordinates,recompute_basis,
-		return_code,size,this_comp_no,xi_offset;
-	FE_value *basis_value,*calculated_value,
-		**component_values,*derivative,*element_value,temp,xi_coordinate;
-	Standard_basis_function *current_standard_basis_function,
-		**component_standard_basis_function;
-	struct FE_field *field;
-	int **component_standard_basis_function_arguments,
-		*current_standard_basis_function_arguments;
-	const Value_storage **component_grid_values_storage,*element_values_storage;
-#if defined (DOUBLE_FOR_DOT_PRODUCT)
-	double sum;
-#else /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-	FE_value sum;
-#endif /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-
-	ENTER(calculate_FE_element_field);
-	return_code=0;
-	FE_value *basis_function_values = 0;
-	if (element_field_values&&xi_coordinates&&values&&
-		(!jacobian||(jacobian&&(element_field_values->derivatives_calculated)))&&
-		(field=element_field_values->field)&&
-		((GENERAL_FE_FIELD != field->fe_field_type)||
-			(basis_function_values=element_field_values->basis_function_values)))
-	{
-		const int dimension = element_field_values->element->getDimension();
-		if ((0<=component_number)&&(component_number<field->number_of_components))
-		{
-			comp_no=component_number;
-			components_to_calculate=1;
-		}
-		else
-		{
-			comp_no=0;
-			components_to_calculate=field->number_of_components;
-		}
-		switch (field->fe_field_type)
-		{
-			case CONSTANT_FE_FIELD:
-			{
-				return_code=1;
-				for (i=0;(i<components_to_calculate)&&return_code;i++)
-				{
-					if (!get_FE_field_FE_value_value(field,comp_no,&values[i]))
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field.  "
-							"Could not get values for constant field %s",field->name);
-						return_code=0;
-					}
-					comp_no++;
-				}
-				if (jacobian)
-				{
-					/* derivatives are zero for constant fields */
-					derivative=jacobian;
-					for (i = (field->number_of_components)*dimension; 0 < i ; --i)
-					{
-						*derivative = 0.0;
-						derivative++;
-					}
-				}
-			} break;
-			case INDEXED_FE_FIELD:
-			{
-				int index,value_no;
-
-				REACCESS(FE_field)(&(element_field_values->field),field->indexer_field);
-				if (calculate_FE_element_field_int_values(/*component_number*/0,
-					element_field_values,xi_coordinates,&index))
-				{
-					/* index numbers start at 1 */
-					if ((1<=index)&&(index<=field->number_of_indexed_values))
-					{
-						return_code=1;
-						value_no = index-1 + comp_no*field->number_of_indexed_values;
-						for (i=0;(i<components_to_calculate)&&return_code;i++)
-						{
-							if (!get_FE_field_FE_value_value(field,value_no,&values[i]))
-							{
-								display_message(ERROR_MESSAGE,
-									"calculate_FE_element_field.  "
-									"Could not get values for constant field %s",field->name);
-								return_code=0;
-							}
-							value_no += field->number_of_indexed_values;
-						}
-						if (jacobian)
-						{
-							/* derivatives are zero for indexed fields */
-							derivative=jacobian;
-							for (i = (field->number_of_components)*dimension; 0 < i ; --i)
-							{
-								*derivative = 0.0;
-								derivative++;
-							}
-						}
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,"calculate_FE_element_field.  "
-							"Index field %s gave out-of-range index %d in field %s",
-							field->indexer_field->name,index,field->name);
-					}
-				}
-				else
-				{
-					display_message(ERROR_MESSAGE,"calculate_FE_element_field.  "
-						"Could not calculate index field %s for field %s at %d-D element %",
-						field->indexer_field->name,field->name,
-						element_field_values->element->getDimension(),
-						element_field_values->element->getIdentifier());
-				}
-				REACCESS(FE_field)(&(element_field_values->field),field);
-			} break;
-			case GENERAL_FE_FIELD:
-			{
-				/* calculate the value for the element field */
-				return_code=1;
-				/* calculate a value for each component */
-				current_standard_basis_function=(Standard_basis_function *)NULL;
-				current_standard_basis_function_arguments=(int *)NULL;
-				component_number_of_values=
-					element_field_values->component_number_of_values;
-				component_values=element_field_values->component_values;
-				component_standard_basis_function=
-					element_field_values->component_standard_basis_functions;
-				component_standard_basis_function_arguments=
-					element_field_values->component_standard_basis_function_arguments;
-				calculated_value=values;
-				if (element_field_values->derivatives_calculated)
-				{
-					derivative=jacobian;
-				}
-				else
-				{
-					derivative=(FE_value *)NULL;
-				}
-				component_number_of_values += comp_no;
-				component_values += comp_no;
-				component_standard_basis_function += comp_no;
-				component_standard_basis_function_arguments += comp_no;
-				number_of_xi_coordinates = dimension;
-				int *element_value_offsets = 0;
-				int *element_value_offset = 0;
-				int number_of_values = 0;
-				int offset = 0;
-				for (cn=0;(cn<components_to_calculate)&&return_code;cn++)
-				{
-					this_comp_no = comp_no + cn;
-					number_in_xi = element_field_values->component_number_in_xi[this_comp_no];
-					if (number_in_xi)
-					{
-						/* grid based */
-						current_standard_basis_function = NULL;
-						current_standard_basis_function_arguments = NULL;
-						/* optimisation: reuse basis from last component if same number_in_xi */
-						recompute_basis = 1;
-						if ((cn > 0) && (last_number_in_xi = element_field_values->component_number_in_xi[this_comp_no - 1]))
-						{
-							recompute_basis = 0;
-							for (i = 0; i < number_of_xi_coordinates; i++)
-							{
-								if (number_in_xi[i] != last_number_in_xi[i])
-								{
-									recompute_basis = 1;
-									break;
-								}
-							}
-						}
-						if (recompute_basis)
-						{
-							number_of_values=1;
-							for (i=number_of_xi_coordinates;i>0;i--)
-							{
-								number_of_values *= 2;
-							}
-							return_code=1;
-							i=0;
-							offset=element_field_values->component_base_grid_offset[this_comp_no];
-							*basis_function_values=1;
-							element_value_offsets = element_field_values->element_value_offsets;
-							*element_value_offsets=0;
-							m=1;
-							while (return_code&&(i<number_of_xi_coordinates))
-							{
-								xi_coordinate=xi_coordinates[i];
-								if (0.0 > xi_coordinate)
-								{
-									xi_coordinate = 0.0;
-								}
-								if (xi_coordinate > 1.0)
-								{
-									xi_coordinate = 1.0;
-								}
-								/* get xi_offset = lower grid number for cell in xi_coordinate
-									 i, and xi_coordinate = fractional xi value in grid cell */
-								if (1.0 == xi_coordinate)
-								{
-									if (number_in_xi[i] > 0)
-									{
-										xi_offset=number_in_xi[i]-1;
-									}
-									else
-									{
-										xi_offset=0;
-									}
-								}
-								else
-								{
-									xi_coordinate *= (FE_value)(number_in_xi[i]);
-									xi_offset=(int)floor((double)xi_coordinate);
-									xi_coordinate -= (FE_value)xi_offset;
-								}
-								offset += xi_offset*element_field_values->component_grid_offset_in_xi[this_comp_no][i];
-								/* add grid_offset in xi_coordinate i for neighbouring grid
-									 points around the linear cell */
-								element_value_offset=element_value_offsets;
-								for (l=m;l>0;l--)
-								{
-									element_value_offset[m]=(*element_value_offset)+
-										element_field_values->component_grid_offset_in_xi[this_comp_no][i];
-									element_value_offset++;
-								}
-								temp = 1.0 - xi_coordinate;
-								basis_value=basis_function_values;
-								if (derivative)
-								{
-									// derivatives are w.r.t. element xi, not the grid element xi
-									FE_value grid_xi_scaling = static_cast<FE_value>(number_in_xi[i]);
-									for (j=1;j<=i;j++)
-									{
-										basis_value=basis_function_values+(j*number_of_values+m);
-										for (l=m;l>0;l--)
-										{
-											basis_value--;
-											basis_value[m]=(*basis_value)*xi_coordinate;
-											*basis_value *= temp;
-										}
-									}
-									j=(i+1)*number_of_values;
-									basis_value=basis_function_values+m;
-									for (l=m;l>0;l--)
-									{
-										basis_value--;
-										basis_value[j]= -(*basis_value)*grid_xi_scaling;
-										basis_value[j+m]= (*basis_value)*grid_xi_scaling;
-									}
-								}
-								basis_value=basis_function_values+m;
-								for (l=m;l>0;l--)
-								{
-									basis_value--;
-									basis_value[m]=(*basis_value)*xi_coordinate;
-									*basis_value *= temp;
-								}
-								m *= 2;
-								i++;
-							}
-						}
-						if (return_code)
-						{
-							size=get_Value_storage_size(field->value_type,
-								(struct FE_time_sequence *)NULL);
-							component_grid_values_storage=
-								element_field_values->component_grid_values_storage + this_comp_no;
-
-							element_values_storage=
-								(*component_grid_values_storage)+size*offset;
-							basis_value=basis_function_values;
-							sum=0;
-							element_value_offset=element_value_offsets;
-							for (j=number_of_values;j>0;j--)
-							{
-#if defined (DOUBLE_FOR_DOT_PRODUCT)
-								sum += (double)(*basis_value)*(double)(*((FE_value *)(
-									element_values_storage+size*(*element_value_offset))));
-#else /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-								sum += (*basis_value)*(*((FE_value *)(
-									element_values_storage+size*(*element_value_offset))));
-#endif /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-								element_value_offset++;
-								basis_value++;
-							}
-							*calculated_value=(FE_value)sum;
-							if (derivative)
-							{
-								for (k=number_of_xi_coordinates;k>0;k--)
-								{
-									sum=0;
-									element_value_offset=element_value_offsets;
-									for (j=number_of_values;j>0;j--)
-									{
-#if defined (DOUBLE_FOR_DOT_PRODUCT)
-										sum += (double)(*basis_value)*(double)(*((FE_value *)(
-											element_values_storage+size*(*element_value_offset))));
-#else /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-										sum += (*basis_value)*(*((FE_value *)(
-											element_values_storage+size*(*element_value_offset))));
-#endif /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-										element_value_offset++;
-										basis_value++;
-									}
-									*derivative=(FE_value)sum;
-									derivative++;
-								}
-							}
-						}
-					}
-					else
-					{
-						/* standard interpolation */
-						/* save calculation when all components use the same basis */
-						/*???DB.  Only good when consecutive components have the same basis.
-							Can do better ? */
-						if ((*component_standard_basis_function!=
-							current_standard_basis_function)||
-							(*component_standard_basis_function_arguments!=
-								current_standard_basis_function_arguments))
-						{
-							current_standard_basis_function=
-								*component_standard_basis_function;
-							current_standard_basis_function_arguments=
-								*component_standard_basis_function_arguments;
-							number_of_values= *component_number_of_values;
-							/* calculate the values for the standard basis functions */
-							if (!(current_standard_basis_function)(
-								current_standard_basis_function_arguments,xi_coordinates,
-								basis_function_values))
-							{
-								display_message(ERROR_MESSAGE,"calculate_FE_element_field.  "
-									"Error calculating standard basis");
-								return_code=0;
-							}
-						}
-						/* calculate the element field value as a dot product of the element
-							 values and the basis function values */
-						basis_value=basis_function_values;
-						element_value= *component_values;
-						sum=0;
-						for (j=number_of_values;j>0;j--)
-						{
-#if defined (DOUBLE_FOR_DOT_PRODUCT)
-							sum += (double)(*element_value)*(double)(*basis_value);
-#else /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-							sum += (*element_value)*(*basis_value);
-#endif /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-							basis_value++;
-							element_value++;
-						}
-						*calculated_value=(FE_value)sum;
-						if (derivative)
-						{
-							for (k=number_of_xi_coordinates;k>0;k--)
-							{
-								sum=0;
-								basis_value=basis_function_values;
-								for (j=number_of_values;j>0;j--)
-								{
-#if defined (DOUBLE_FOR_DOT_PRODUCT)
-									sum += (double)(*element_value)*(double)(*basis_value);
-#else /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-									sum += (*element_value)*(*basis_value);
-#endif /* defined (DOUBLE_FOR_DOT_PRODUCT) */
-									basis_value++;
-									element_value++;
-								}
-								*derivative=(FE_value)sum;
-								derivative++;
-							}
-						}
-					}
-					component_number_of_values++;
-					component_values++;
-					component_standard_basis_function++;
-					component_standard_basis_function_arguments++;
-					calculated_value++;
-				}
-			} break;
-			default:
-			{
-				display_message(ERROR_MESSAGE,
-					"calculate_FE_element_field.  Unknown field type");
-			} break;
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"calculate_FE_element_field.  Invalid argument(s)\n"
-			"element_field_values %p, xi_coordinates %p, values %p, jacobian %p",
-			element_field_values,xi_coordinates,values,jacobian);
-	}
-	LEAVE;
-
-	return (return_code);
-} /* calculate_FE_element_field */
-
-int calculate_FE_element_field_as_string(int component_number,
-	struct FE_element_field_values *element_field_values,
-	const FE_value *xi_coordinates, char **string)
-/*******************************************************************************
-LAST MODIFIED : 19 October 1999
-
-DESCRIPTION :
-Calculates the values of element field specified by the <element_field_values>
-at the <xi_coordinates> and returns them as the allocated <string>. Only the
-<component_number>+1 component will be calculated if
-0<=component_number<number of components. If more than 1 component is calculated
-then values are comma separated. Derivatives are not included in the string,
-even if calculated for the <element_field_values>.
-It is up to the calling function to DEALLOCATE the returned string.
-==============================================================================*/
-{
-	char temp_string[40];
-	int components_to_calculate,error,i,return_code;
-	struct FE_field *field;
-
-	ENTER(calculate_FE_element_field_as_string);
-	return_code=0;
-	(*string)=(char *)NULL;
-	if (element_field_values&&xi_coordinates&&string&&
-		(field=element_field_values->field))
-	{
-		if ((0<=component_number)&&(component_number<field->number_of_components))
-		{
-			components_to_calculate=1;
-		}
-		else
-		{
-			components_to_calculate=field->number_of_components;
-		}
-		switch (field->value_type)
-		{
-			case FE_VALUE_VALUE:
-			{
-				FE_value *values;
-
-				if (ALLOCATE(values,FE_value,components_to_calculate))
-				{
-					if (calculate_FE_element_field(component_number,
-						element_field_values,xi_coordinates,values,
-						/*jacobian*/(FE_value *)NULL))
-					{
-						error=0;
-						for (i=0;i<components_to_calculate;i++)
-						{
-							if (0<i)
-							{
-								sprintf(temp_string,",%g",values[i]);
-							}
-							else
-							{
-								sprintf(temp_string,"%g",values[i]);
-							}
-							append_string(string,temp_string,&error);
-						}
-						return_code= !error;
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_as_string.  "
-							"Could not calculate FE_value values");
-					}
-					DEALLOCATE(values);
-				}
-			} break;
-			case INT_VALUE:
-			{
-				int *values;
-
-				if (ALLOCATE(values,int,components_to_calculate))
-				{
-					if (calculate_FE_element_field_int_values(component_number,
-						element_field_values,xi_coordinates,values))
-					{
-						error=0;
-						for (i=0;i<components_to_calculate;i++)
-						{
-							if (0<i)
-							{
-								sprintf(temp_string,",%d",values[i]);
-							}
-							else
-							{
-								sprintf(temp_string,"%d",values[i]);
-							}
-							append_string(string,temp_string,&error);
-						}
-						return_code= !error;
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_as_string.  "
-							"Could not calculate int values");
-					}
-					DEALLOCATE(values);
-				}
-			} break;
-			case STRING_VALUE:
-			{
-				char **values;
-
-				if (ALLOCATE(values,char *,components_to_calculate))
-				{
-					if (calculate_FE_element_field_string_values(component_number,
-						element_field_values,xi_coordinates,values))
-					{
-						error=0;
-						for (i=0;i<components_to_calculate;i++)
-						{
-							if (0<i)
-							{
-								append_string(string,",",&error);
-							}
-							append_string(string,values[i],&error);
-						}
-						for (i=0;i<components_to_calculate;i++)
-						{
-							DEALLOCATE(values[i]);
-						}
-						return_code= !error;
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_as_string.  "
-							"Could not calculate string values");
-					}
-					DEALLOCATE(values);
-				}
-			} break;
-			default:
-			{
-				display_message(ERROR_MESSAGE,
-					"calculate_FE_element_field_as_string.  Unknown value type %s",
-					Value_type_string(field->value_type));
-			} break;
-		}
-		if (!return_code)
-		{
-			display_message(ERROR_MESSAGE,
-				"calculate_FE_element_field_as_string.  Failed");
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"calculate_FE_element_field_as_string.  Invalid argument(s)");
-	}
-	LEAVE;
-
-	return (return_code);
-} /* calculate_FE_element_field_as_string */
-
-int calculate_FE_element_field_int_values(int component_number,
-	struct FE_element_field_values *element_field_values,
-	const FE_value *xi_coordinates, int *values)
-/*******************************************************************************
-LAST MODIFIED : 19 October 1999
-
-DESCRIPTION :
-Calculates the <values> of the integer field specified by the
-<element_field_values> at the <xi_coordinates>. The storage for the <values>
-should have been allocated outside the function. Only the <component_number>+1
-component will be calculated if 0<=component_number<number of components. For a
-single component, the value will be put in the first position of <values>.
-==============================================================================*/
-{
-	int cn,comp_no,components_to_calculate,*element_int_values,i,*number_in_xi,
-		offset,return_code,this_comp_no,xi_offset;
-	FE_value xi_coordinate;
-	struct FE_field *field;
-
-	ENTER(calculate_FE_element_field_int_values);
-	return_code=0;
-	if (element_field_values&&xi_coordinates&&values&&
-		(field=element_field_values->field)&&(INT_VALUE==field->value_type))
-	{
-		if ((0<=component_number)&&(component_number<field->number_of_components))
-		{
-			comp_no=component_number;
-			components_to_calculate=1;
-		}
-		else
-		{
-			comp_no=0;
-			components_to_calculate=field->number_of_components;
-		}
-		switch (field->fe_field_type)
-		{
-			case CONSTANT_FE_FIELD:
-			{
-				return_code=1;
-				for (i=0;(i<components_to_calculate)&&return_code;i++)
-				{
-					if (!get_FE_field_int_value(field,comp_no,&values[i]))
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_int_values.  "
-							"Could not get values for constant field %s",field->name);
-						return_code=0;
-					}
-					comp_no++;
-				}
-			} break;
-			case INDEXED_FE_FIELD:
-			{
-				int index,value_no;
-
-				REACCESS(FE_field)(&(element_field_values->field),field->indexer_field);
-				if (calculate_FE_element_field_int_values(/*component_number*/0,
-					element_field_values,xi_coordinates,&index))
-				{
-					/* index numbers start at 1 */
-					if ((1<=index)&&(index<=field->number_of_indexed_values))
-					{
-						return_code=1;
-						value_no = index-1 + comp_no*field->number_of_indexed_values;
-						for (i=0;(i<components_to_calculate)&&return_code;i++)
-						{
-							if (!get_FE_field_int_value(field,value_no,&values[i]))
-							{
-								display_message(ERROR_MESSAGE,
-									"calculate_FE_element_field_int_values.  "
-									"Could not get values for constant field %s",field->name);
-								return_code=0;
-							}
-							value_no += field->number_of_indexed_values;
-						}
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_int_values.  "
-							"Index field %s gave out-of-range index %d in field %s",
-							field->indexer_field->name,index,field->name);
-					}
-				}
-				else
-				{
-					display_message(ERROR_MESSAGE,
-						"calculate_FE_element_field_int_values.  "
-						"Could not calculate index field %s for field %s at %d-D element %",
-						field->indexer_field->name,field->name,
-						element_field_values->element->getDimension(),
-						element_field_values->element->getIdentifier());
-				}
-				REACCESS(FE_field)(&(element_field_values->field),field);
-			} break;
-			case GENERAL_FE_FIELD:
-			{
-				return_code = 1;
-				const int number_of_xi_coordinates = element_field_values->element->getDimension();
-				for (cn=0;(cn<components_to_calculate)&&return_code;cn++)
-				{
-					this_comp_no = comp_no + cn;
-					number_in_xi = element_field_values->component_number_in_xi[this_comp_no];
-					if (number_in_xi)
-					{
-						/* for integer, get value at nearest grid point */
-						return_code=1;
-						offset=element_field_values->component_base_grid_offset[this_comp_no];
-						for (i=0;(i<number_of_xi_coordinates)&&return_code;i++)
-						{
-							xi_coordinate=xi_coordinates[i];
-							if (xi_coordinate < 0.0)
-							{
-								xi_coordinate = 0.0;
-							}
-							else if (xi_coordinate > 1.0)
-							{
-								xi_coordinate = 1.0;
-							}
-							/* get nearest xi_offset */
-							xi_offset=(int)floor((double)number_in_xi[i]*(double)xi_coordinate+0.5);
-							offset += xi_offset*element_field_values->component_grid_offset_in_xi[this_comp_no][i];
-						}
-						element_int_values = (int *)element_field_values->component_grid_values_storage[this_comp_no];
-						values[cn] = element_int_values[offset];
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_int_values.  "
-							"Non-grid-based component for integer valued field");
-						return_code = 0;
-					}
-				}
-			} break;
-			default:
-			{
-				display_message(ERROR_MESSAGE,
-					"calculate_FE_element_field_int_values.  Unknown field type");
-			} break;
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"calculate_FE_element_field_int_values.  Invalid argument(s)");
-	}
-	LEAVE;
-
-	return (return_code);
-} /* calculate_FE_element_field_int_values */
-
-int calculate_FE_element_field_string_values(int component_number,
-	struct FE_element_field_values *element_field_values,
-	const FE_value *xi_coordinates, char **values)
-/*******************************************************************************
-LAST MODIFIED : 19 October 1999
-
-DESCRIPTION :
-Returns allocated copies of the string values of the field specified by the
-<element_field_values> at the <xi_coordinates>. <values> must be allocated with
-enough space for the number_of_components strings, but the strings themselves
-are allocated here. Only the <component_number>+1 component will be calculated
-if 0<=component_number<number of components. For a single component, the value
-will be put in the first position of <values>.
-It is up to the calling function to deallocate the returned string values.
-==============================================================================*/
-{
-	int comp_no,components_to_calculate,i,j,return_code;
-	struct FE_field *field;
-
-	ENTER(calculate_FE_element_field_string_values);
-	return_code=0;
-	if (element_field_values&&xi_coordinates&&values&&
-		(field=element_field_values->field)&&(STRING_VALUE==field->value_type))
-	{
-		if ((0<=component_number)&&(component_number<field->number_of_components))
-		{
-			comp_no=component_number;
-			components_to_calculate=1;
-		}
-		else
-		{
-			comp_no=0;
-			components_to_calculate=field->number_of_components;
-		}
-		switch (field->fe_field_type)
-		{
-			case CONSTANT_FE_FIELD:
-			{
-				return_code=1;
-				for (i=0;(i<components_to_calculate)&&return_code;i++)
-				{
-					if (!get_FE_field_string_value(field,comp_no,&values[i]))
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_string_values.  "
-							"Could not get values for constant field %s",field->name);
-						return_code=0;
-						for (j=0;j<i;j++)
-						{
-							DEALLOCATE(values[j]);
-						}
-					}
-					comp_no++;
-				}
-			} break;
-			case INDEXED_FE_FIELD:
-			{
-				int index,value_no;
-
-				REACCESS(FE_field)(&(element_field_values->field),field->indexer_field);
-				if (calculate_FE_element_field_int_values(/*component_number*/0,
-					element_field_values,xi_coordinates,&index))
-				{
-					/* index numbers start at 1 */
-					if ((1<=index)&&(index<=field->number_of_indexed_values))
-					{
-						return_code=1;
-						value_no = index-1 + comp_no*field->number_of_indexed_values;
-						for (i=0;(i<components_to_calculate)&&return_code;i++)
-						{
-							if (!get_FE_field_string_value(field,value_no,&values[i]))
-							{
-								display_message(ERROR_MESSAGE,
-									"calculate_FE_element_field_string_values.  "
-									"Could not get values for constant field %s",field->name);
-								return_code=0;
-								for (j=0;j<i;j++)
-								{
-									DEALLOCATE(values[j]);
-								}
-							}
-							value_no += field->number_of_indexed_values;
-						}
-					}
-					else
-					{
-						display_message(ERROR_MESSAGE,
-							"calculate_FE_element_field_string_values.  "
-							"Index field %s gave out-of-range index %d in field %s",
-							field->indexer_field->name,index,field->name);
-					}
-				}
-				else
-				{
-					display_message(ERROR_MESSAGE,
-						"calculate_FE_element_field_string_values.  "
-						"Could not calculate index field %s for field %s at %d-D element %",
-						field->indexer_field->name,field->name,
-						element_field_values->element->getDimension(),
-						element_field_values->element->getIdentifier());
-				}
-				REACCESS(FE_field)(&(element_field_values->field),field);
-			} break;
-			case GENERAL_FE_FIELD:
-			{
-				display_message(ERROR_MESSAGE,
-					"calculate_FE_element_field_string_values.  "
-					"General fields not supported");
-			} break;
-			default:
-			{
-				display_message(ERROR_MESSAGE,
-					"calculate_FE_element_field_string_values.  Unknown field type");
-			} break;
-		}
-	}
-	else
-	{
-		display_message(ERROR_MESSAGE,
-			"calculate_FE_element_field_string_values.  Invalid argument(s)");
-	}
-	LEAVE;
-
-	return (return_code);
-} /* calculate_FE_element_field_string_values */
 
 bool equivalent_FE_field_in_elements(struct FE_field *field,
 	struct FE_element *element_1, struct FE_element *element_2)
@@ -8940,7 +6870,7 @@ public:
 
 template <typename ObjectType> class SelfDestruct
 {
-	typedef int (DestroyFunction)(ObjectType* objectAddress);
+	typedef int (DestroyFunction)(ObjectType &objectAddress);
 	ObjectType& object;
 	DestroyFunction *destroyFunction;
 
@@ -8953,7 +6883,7 @@ public:
 
 	~SelfDestruct()
 	{
-		(this->destroyFunction)(&object);
+		(this->destroyFunction)(object);
 	}
 };
 
@@ -9000,14 +6930,13 @@ bool FE_element_smooth_FE_field(struct FE_element *element,
 		(3 == dimension) ? 8 :
 		0;
 	FE_value component_value[8], xi[MAXIMUM_ELEMENT_XI_DIMENSIONS];
-	FE_element_field_values *fe_element_field_values = CREATE(FE_element_field_values)();
-	SelfDestruct<FE_element_field_values *> sd1(fe_element_field_values, DESTROY(FE_element_field_values));
+	FE_element_field_evaluation *fe_element_field_evaluation = FE_element_field_evaluation::create();
+	SelfDestruct<FE_element_field_evaluation *> sd1(fe_element_field_evaluation, FE_element_field_evaluation::deaccess);
 	// need to calculate field values to evaluate field in element corners,
 	// ensuring optional modify function, version mapping etc. applied
 	int return_code = 1;
-	if (!calculate_FE_element_field_values(element, fe_field, time,
-		/*calculate_derivatives*/0, fe_element_field_values,
-		/*top_level_element*/(struct FE_element *)NULL))
+	if (!fe_element_field_evaluation->calculate_values(fe_field, element,
+		time, /*calculate_derivatives*/0, /*top_level_element*/0))
 	{
 		display_message(ERROR_MESSAGE,
 			"FE_element_smooth_FE_field.  Could not calculate element field values");
@@ -9080,9 +7009,8 @@ bool FE_element_smooth_FE_field(struct FE_element *element,
 			xi[0] = (FE_value)(n & 1);
 			xi[1] = (FE_value)((n & 2)/2);
 			xi[2] = (FE_value)((n & 4)/4);
-			if (!calculate_FE_element_field(componentNumber,
-				fe_element_field_values, xi, &(component_value[n]),
-				/*jacobian*/(FE_value *)NULL))
+			if (!fe_element_field_evaluation->evaluate_real(componentNumber,
+				xi, &(component_value[n]), /*jacobian*/0))
 			{
 				display_message(ERROR_MESSAGE,
 					"FE_element_smooth_FE_field.  Could not calculate element field");
